Enforce architecture boundaries and legacy shim governance
